requirements.txt
pyserial==3.5
pygame==2.5.2
numpy==1.26.4

